Add ISO 639 languages fixtures and import command
ISO 639 languages are included in a CSV file for convenient reading and
editing. Data source is from the US Library of Congress
(http://loc.gov/standards/iso639-2/php/code_list.php). French and German
names are included for completeness.

The management command loads the fixture file (or a different file) and
runs through the list ensuring that it does not duplicate any tags. Each
tag is categorized and assigned a creator and updater, which can be
overridden in the command arguments.

By default this only loads lanugages with ISO 639-1 codes, excluding
many languages. The complete list includes many listing which should
probably _not_ be included (e.g. Klingon), and without a strategy for
excluding these the command and data combination should be more
conservative.

Closes gh-232
[Fixes #116086691]