Add data fixture and command to load countries/country codes
Similar to the load_orb_languages command, this combines a default
fixture file with a command that conditionally loads data. It allows for
configuration against:

- Default tag owner user
- Fixture file (CSV)
- Name of the CSV header for the country name
- Name of the CSV header for the country code

Missing tags are added with a default category of "Geography" and the
default geography tag image. These are not configurable.

The data is sourced from
http://data.okfn.org/data/core/country-codes/r/country-codes.csv under
public domain license (subject to original ISO licensing).

Closes gh-231
[#116086693]