Send email on final resource approval/rejection
Involves removing criteria from final email as this is not collected
from the staff reviewer, only notes.

Notes are pulled from *possibly* multiple resource workflows. This
should really never be the case, but the queryset construction avoids
issues dealing with multiple relations inherent in the one-to-many
nature of the ForeignKey.

Closes gh-407