Ensure tag names are checked against local lang and default field
This prevents the scenario where the lookup for an existing tag returns nothing
beacuse the tag name, present in the default name field, is not returned for the
current language translation. E.g. a tag was added in English as "Medicine"
which adds "Medicine" to both the `name` and `name_en` database fields.

If then someone is using the site in Spanish, the modeltranslation query will
automatically query *only* the `name_es` field. Finding nothing, an attempt is
made to create a tag named "Medicine", which puts "Medicine" into both the
`name_es` and `name` fields. Since both fields have unique constraints, this
fails on the `name` field.

The update here checks both the default field and current translated field.

Fixes gh-458